+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.91] - Not yet released, expected end of May 2025
-- TODO: add items as needed
+ - Fix generated reader auth certificates to contain required extensions from ISO/IEC 18013-5:2021 Annex B.1.7.
 
 ## [0.90] - 2025-03-26
 This is the first release with the CHANGELOG.md file and also the first release using the new